Rename SquirlsConfiguration to Squirls
